41r/pyproject.toml
[project]
name = "41r"
version = "0.1.0"
description = "41R Persona Market — Predictive A/B testing with calibrated personas"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "playwright>=1.49.0",
    "weasyprint>=62.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100
